personal_website/backend/requirements.txt
# Core dependencies
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
python-slugify==8.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
PyJWT==2.8.0
Flask-Login==0.6.2
Pillow==10.0.0
bleach==6.0.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy-Utils==0.41.1

# Forms and validation
WTForms==3.0.1
email-validator==2.0.0

# Authentication and security
bcrypt==4.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2

# File uploads
Flask-Uploads==0.2.1

# Error handling
sentry-sdk==1.30.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
tqdm==4.66.1

# Development dependencies
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
pre-commit==3.4.0
Flask-DebugToolbar==0.13.1

# Production dependencies
gunicorn==21.2.0
Flask-SSLify==0.1.5
Werkzeug-ProxyFix==1.0.4
waitress==2.1.2